Moved AudioKit Safety extension to its own file
